Code/requirements.txt
# Django Core
Django==5.2.8

# Database
# Using SQLite (default) - no additional driver needed for development
# For production with PostgreSQL, add: psycopg2-binary==2.9.9

# Excel Processing
openpyxl==3.1.2
pandas==2.1.4

# Data Processing
numpy==1.26.2

# Environment Variables
python-decouple==3.8

# Security
python-dotenv==1.0.0

# File Handling
Pillow==10.1.0

# Web Server (Production)
gunicorn==21.2.0
whitenoise==6.6.0

# Development Tools
ipython==8.18.1

# Async Task Processing (Optional - for future implementation)
# celery==5.3.4
# redis==5.0.1
# django-celery-beat==2.5.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Code Quality
flake8==6.1.0
black==23.12.0

# API Development (Optional - for future API)
# djangorestframework==3.14.0
# django-cors-headers==4.3.1

# Additional Django Packages
django-crispy-forms==2.1
crispy-bootstrap5==2.0.0

# Monitoring and Logging
# sentry-sdk==1.39.1  # Uncomment for error tracking in production
